Skip testing subnormals when flushed to zero
When flush-to-zero (FTZ) mode is enabled, std::nextafter() returns zero
when a subnormal gets passed in. This can cause infinite loops.

The inc() function returns the next larger value which does not compare
equal to the input. The only exception is negative zero, which produces
positive zero, since the CPU still discerns between these two when FTZ
has been enabled.

Enabling FTZ was not found to have a significant effect on the run time
of SinExhaustive though (~80 seconds on an Intel Core i9-8950HK).
According to https://stackoverflow.com/questions/60969892, there is only
a penalty when multiplying numbers which would produce a subnormal
result.

Bug: b/169904252
Change-Id: Idff987a4c14bd4f778e940ec4fc4de26e0a74372