Skip testing subnormals when flushed to zero
When flush-to-zero (FTZ) mode is enabled, std::nextafter() returns zero
when a subnormal gets passed in. This can cause infinite loops when
trying to iterate over a range.

The inc() function returns the next larger value which does not compare
equal to the input. The only exception is negative zero, which produces
positive zero, since the CPU still discerns between these two when FTZ
has been enabled.

Bug: b/169904252
Change-Id: Idff987a4c14bd4f778e940ec4fc4de26e0a74372