Fix exp2 conformance
A 1-bit difference in the coefficients drops the ULP error over the
-10..+10 range to 2.90, keeping within a 0.927 margin of Vulkan's high-
precision tolerance.

Note that theoretically the previous polynomial coefficients should also
produce results with an error below 3 ULP, but at x values around -0.067
the polynomial is actually evaluated at 0.933 which is both near a peak
in the graph and where small rounding errors become relatively large.
Also, the |x| portion in the 2 + 3 * |x| tolerance formula doesn't help
us out in this area.

This adjusted coefficient does not produce the best possible result, but
this change is just meant to establish a baseline implementation which
is 100% conformant.

Bug: b/169754022
Change-Id: I9fdeee304f6790b0919578926f82b20c0176fe58