Use floor(x) to compute the fractional part of x
Previously we used round(x - 0.5) as part of computing the fractional
part of x for the exp2(x) algorithm, because single-instruction floor()
operations weren't available until SSE4.1. Now it's ubiquitous and we
can eliminate the subtraction. This didn't alter the maximum ULP error.

Exp2() was also refactored to separate the polynomial coefficients from
the polynomial evaluation.

Bug: b/169754022
Change-Id: I77790ec5624932e13b072741c7e1d095fbb83f8f