Don't treat all warnings as errors in tests and benchmarks
Adding tests should be low friction and compiler warnings are
exceedingly rarely helpful to write better tests.

For example one may use variables which are assigned but not used,
which are useful during interactive debugging or as part of an assert,
but trigger the unused-but-set-variable warning which we treat as an
error. While we could decorate the variable with [[maybe_unused]] this
is distracting while reading the code.

Also, warnings differ between compilers and may only appear in presubmit
tests, further delaying the development of new tests.

That said, most tests still use ${SWIFTSHADER_COMPILE_OPTIONS}, which
adds some specific warnings treated as error as well as other options
considered generally useful. The Vulkan tests and benchmarks don't use
it because it requires disabling exceptions support, which makes the
entirely vulkan.hpp API change and would require us to rewrite things.

Bug: b/158231104
Change-Id: I445e3729d4d0500893326f7f802aeca171855612