Support render area instead of extent in Renderer::draw()
Some Vulkan 1.3 features related to dynamic rendering expect the
render area to support an offset, as opposed to only requiring a
width and height, so this CL adds support for a render area in
Renderer::draw().

Note: This may be related to b/177842474, but since we're
missing test coverage for it, we can't test it yet, but
this CL might be a step in the right direction.

Bug: b/204502119
Bug: b/177842474
Change-Id: I6d27f353279ef6d46f86c8c65c8e0d1fa76ff927