Eliminate initialization of fragment shader inputs
The Vulkan spec on Shader Interface Matching [0] states that:
"Any input value to a shader stage is well-defined as long as the
preceding stages writes to a matching output, as described above."

The values are therefore undefined if they weren't written by the
previous stage (i.e. the vertex shader). To detect uses of uninitialized
values with MemorySanitizer we must omit initializing them ourselves.

This is also a minor optimization.

[0] https://www.khronos.org/registry/vulkan/specs/1.3/html/vkspec.html#interfaces-iointerfaces-matching

Bug: b/155148722
Bug: b/219942708
Bug: b/190063459
Change-Id: I6018278585ee7e61069448271a83ee568028e690