Use legacy math precision for the Pipeline source set as well
Previously SWIFTSHADER_LEGACY_TRANSCENDENTALS was only enabled on the
libvulkan target, which doesn't affect the Pipeline source set
dependency. This change sets it explicitly for the latter so it is
observed by the src/Pipeline/ShaderCore.cpp file where it's needed.

Also the macro has been renamed to SWIFTSHADER_LEGACY_PRECISION since
we might need it for more than just transcendentals, but will always be
precision related. For example use of MulAdd() may use FMA operations
which can improve precision.

Bug: chromium:1299047
Change-Id: I6c99b17c94ff96ea264c0cb4a7c36819de446c39