Make inc() flush subnormals to zero when DAZ is enabled
Previously the smallest negative normalized floating-point value would
become a negative subnormal when calling inc() on it, regardless of
whether denormals-are-zero is enabled. Calling inc() again on that
result would then produce the smallest positive normalized value.
This change instead flushes the subnormal to zero if DAZ is enabled.

This is useful for when we're iterating over a range and want to know
whether the iterator is an integer. A single-precision subnormal would
compare equal to zero, but when cast to a double would not longer be
subnormal and thus produce a result different from when zero is used as
input.

Note that this change entangles DAZ and FTZ, in the sense that FTZ
behavior is achieved when DAZ is enabled. Thus we expect either both to
be enabled, or neither.

Bug: b/169904252
Change-Id: I1204466cd1793cb9e3011c3549f672faf25d9ddc