Implement separate DAZ/FTZ control for Windows
Flush-To-Zero (FTZ) controls whether subnormal results of floating-point
operations are replaced by zero. Denormals-Are-Zero (DAZ) controls
whether subnormal operands of floating-point operations are treated as
zero.

The Windows _controlfp() function does not offer a way to set each of
them separately, but defines four constants that represent the four
possible combinations. By first retrieving the current status of the
control word to check whether the other mode is on or off, we can
selectively enable or disable each mode.

Bug: b/169904252
Change-Id: I895220ff7c980e5b0ddc0b8a62649a8c32b99d73