Refactor image offset methods
Previously there were three Image::getMemoryOffset() methods. One was
being used for obtaining the subresource offset, so it has been
renamed to getSubresourceOffset(). It also gets called by
getTexelPointer(), which expects it to include the device memory
binding offset. This difference in usage is at the root of a bug but
this change doesn't address that so it's a pure refactoring.

The method now called getSubresourceOffset() previously called another
getMemoryOffset() method for obtaining the  mipmap level offset, which
has now been merged into getSubresourceOffset() since it was small and
not used elsewhere.

The third getMemoryOffset() was for obtaining the aspect offset, so it
has been renamed to getAspectOffset(). Although it's only called by the
new getSubresourceOffset(), it's a bit larger so it's kept separate.

Bug: b/208505033
Change-Id: Ic9f48e32a91a4ad6fbcc26e0dbb442384ccc6ac6