Fix image subresource offset query
Previously the VkSubresourceLayout::offset included the offset within
the device memory allocation bound to the image. This change removes
that offset, and explicitly adds it for the getTexelPointer() method
which was relying on this previous behavior.

Also, for images with disjoint planes the subresource offset is relative
to the plane containing that subresource, so getSubresourceOffset() now
leaves out the aspect offset for these images to avoid counting it
twice.

Bug: b/208505033
Change-Id: If20f37a00ae24ef244940878e60fd1331b95edcc