Implement implicit broadcast of scalar constants
This change enables syntax such as `v + 0.5f` where `v` is a Float4
vector. The 0.5f is implicitly converted to an RValue<Float4>. Note that
this also works for any C++ float values, including variables, but from
Reactor's perspective it's a constant.

Explicit operators for adding RValue<Int> and RValue<Int4> have been
removed since they caused ambiguity. This was only used by one instance
of adding Int and SIMD::Int, which was changed to use an explicit
broadcast.

Bug: b/221072484
Change-Id: I50ba159b8f62fa3c7198b7dc55a9832ea7cff03c