Remove duplicate transcendental functions
Functions such as exponential() were considered general purpose but we
practically only needed implementations of transcendental and
trigonometric functions for the corresponding shader operations.

power() was used in sRGB to linear color space conversion but has been
replace with the identical Pow() implementation. It should be optimized
using relaxed precision operations or a tailor made approximation.

Bug: b/169755552
Change-Id: I7b0e25b58fda70ac1ca020b9462e80387d1ba0d6