Treat YCbCr images bound to Android hardware buffers as disjoint
Android hardware buffers can hold properties for multiple image planes
through the AHardwareBuffer_Planes structure. This includes per-plane
data pointers to the first pixel of that plane. Thus images with a
multi-planar format bound to an AHB buffer are essentially disjoint even
though they're not created using the VK_IMAGE_CREATE_DISJOINT_BIT flag.

DeviceMemory::hasExternalImageProperties() has been renamed to
hasExternalImagePlanes() to clarify that it's about storing per-plane
information.

Note that AHardwareBufferExternalMemory::externalImageMemoryOffset()
treats the difference in image plane data pointer as an offset. This is
somewhat hacky, but is left for a future refactoring to improve on. The
fact that Image::getMemoryOffset() skips the vkBindImageMemory()
memoryOffset parameter is correct due to the spec requirement that
"Vulkan images bound to memory imported from an Android hardware buffer
must use dedicated allocations", and VUID-vkBindImageMemory-memory-02628
stating that memoryOffset must be zero for dedicated allocations and we
don't support VK_NV_dedicated_allocation_image_aliasing at this time.

Bug: b/208505033
Change-Id: I5989db80b2cda2e237b3a9d579d6ea98c473de67