Fix obtaining the RelaxedPrecision decoration from the result ID
Unlike GLSL which determines the precision of operations based on the
qualifiers of the operands, SPIR-V requires decorating the result of the
instruction: https://www.khronos.org/registry/SPIR-V/specs/unified1/SPIRV.html#RelaxedPrecisionSection

We were accidentally exercising the expected precision in the
dEQP-VK.glsl.builtin.precision.* tests only because the instructions
directly consumed interface variables, which can also be decorated with
RelaxedPrecision to signify that loads and stores may reduce their
precision (while still logically remaining 32-bit). It is incorrect to
use this input parameter decoration to affect the precision of the
instructions operating on them.

Bug: b/222218659
Change-Id: Ic2cedc53fc773269c73bf070248063bb6e932669