Pass the RelaxedPrecision decoration value to all affected instructions
GLSL.std.450 math functions and a few other instructions have lower
precision requirements when the RelaxedPrecision decoration is present:
https://www.khronos.org/registry/vulkan/specs/1.3/html/vkspec.html#spirvenv-precision-operation

This change does not alter any results. It just enables these
instructions to be optimized for relaxed precision in future changes.

The return type of the functions has been changed to an RValue<> to
prevent accidentally assigning values to them.

Bug: b/222218659
Change-Id: I43c51a597e4445e5e686b9cf3120abdf599d8dae