Implement a relaxed precision log2()
When decoared with RelaxedPrecision, the SPIR-V log2() operation only
has to be accurate to 3 ULP of the binary16 floating-point format,
outside of the [0.5, 2] interval, and have an absolute error < 2^-7
inside of this interval. This implementation achieves 0.80 ULP and
0.10 of the absolute error.

It only uses 4 FMA operations, versus 7 for the high-precision version,
and also eliminates the need for a division.

This reduction in instructions translates to a speedup in
PipelineBenchmarks. Measured on a Xeon Gold 6154:

Transcendental1/sw_Log2_highp/10    30.2 ns
Transcendental1/sw_Log2_mediump/10  10.4 ns

Bug: b/222218659
Change-Id: I4b2663b0b24b3bfd01742620d7ee5ab8c6ef4d08