Implement a relaxed precision exp2()
The SPIR-V exp2() operation only has to be accurate to 1 + 2 * |x| ULP
of the binary16 floating-point format when decorated with
RelaxedPrecision. This implementation achieves 0.13 ULP.

It only uses 4 FMA operations, versus 6 for the high-precision version,
and also doesn't require a shift or integer addition (one of the FMA
operations performs the work of moving the radix to the position between
the binary32's exponent and mantissa fields).

This reduction in instructions translates to a speedup in
PipelineBenchmarks (measured on a Xeon Gold 6154):

Transcendental1/sw_Exp2_highp/10    24.4 ns
Transcendental1/sw_Exp2_mediump/10  18.0 ns

Bug: b/169754022
Change-Id: I8a915abb7869dd40d7c2e52dfc2550839868575d