Support specifying math precision through a template argument
The use of a Boolean parameter to select relaxed precision is fine for
functions such as Asin() which are only called from the SpirvShader
code and pass in a verbose d.RelaxedPrecision argument. But functions
like Sqrt() are also used in places like the VK_BLEND_OP_SOFTLIGHT_EXT
implementation, and Pow() is used in sRGB conversion, where there is no
concept of relaxed precision decorations and passing in true or false
would not have good readability.

To create a reasonably elegant syntax the Highp and Mediump enums have
been defined and they can be used as a template argument for these
functions.

The Vulkan spec states for VK_EXT_blend_operation_advanced that the
"blending precision may be limited to 16-bit floating-point", so we can
use Sqrt<Mediump>(). Likewise for sRGB conversion Pow<Mediump>()
suffices since this non-linear fixed-point encoding only makes sense for
relatively low bit width color components.

Bug: b/222218659
Change-Id: Id3bc4fa68f38574ff23b125befbac68072d39ee1