Improve exp2 accuracy
Previously the worst error was 2.95602512 ULP. This change reduces it
to 2.65837669 ULP (2.29829574 with FMA), staying well clear of Vulkan's
3 + 2 * |x| ULP tolerance.

This was achieved by using 1/x as the weight function. Previously
(3+2*x-1)/x was used, but as we're approximating 2^x as f(x) * x + 1
we really only need to take the x factor into account if we just want
the ULP error over the entire range to be minimal. The fact that
Vulkan defines the tolerance as a linearly increasing function doesn't
really allow us to have a bigger error for f(x) near x=1 because that's
where we also lose most precision due to loss of significance in the
polynomial's additions/subtractions.

Bug: b/169754022
Change-Id: Ide599f8e4f054a8d50df9637f19a54a2616676c5