Improve exp2 accuracy through weight adjustment
Previously the worst error was 2.95602512 ULP. This change reduces it
to 2.65837669 ULP (2.29829574 with FMA), staying well clear of Vulkan's
3 + 2 * |x| ULP tolerance.

This was achieved by using 1/x as the weight function for the Remez
algorithm. Previously (3+2*x-1)/x was used, but while the denominator is
essential due to accounting for the use of f(x) * x + 1 to approximate
2^x, the nominator did not match the actual curve of the error
tolerance. ulp(2^x) is a constant in the range [0, 1] due to 2^x only
varying between 1 and 2.

The 3 + 2 * |x| ULP formula adds only a slight slope to that, but
doesn't actually allow us to have a bigger tolerance toward x=1, due to
the polynomial having more loss of significance in that region. Hence
we stick to optimizing for a constant maximum error.

Bug: b/169754022
Change-Id: Ide599f8e4f054a8d50df9637f19a54a2616676c5