Force-inline all SPIR-V function calls
SwiftShader does not support function calls in SPIR-V shader code at the
moment, which isn't a big issue due to running the InlineExhaustivePass
optimization pass. This is part of the default set of optimizations
SPIRV-Tools applies. However, the optimizer respects the DontInline
flag, and leaves calls to functions that declare this flag not inlined.

This change removes the DontInline flags so that functions always get
inlined.

Note this does not violate the spec since the flag is a "Strong request,
to the extent possible, to not inline the function." but not a demand.
Some GPUs are assumed also not to support function calls and must force-
inline every call too.

Bug: b/172212963
Change-Id: I8912df1d99983eb2831f35e413cb0070fa628368