Regres: Include UNSUPPORTED() failures in CI test runs
UNSUPPORTED() is an assert-like macro which indicates the test hits a
path for a feature that is not supported, and isn't advertised as such.
This typically indicates a test bug, so it's worth running these as part
of CI to check when they're fixed, without having to explicitly add them
to the test run.

Only NOT_SUPPORTED tests should be skipped in quick CI test runs since
these are for features previously known by Regres not to be supported by
SwiftShader.

Change-Id: Ife75f095c9f6c6853d6f737a6a143c1e7f140daf