Convert failing assert condition
In the class FrameBufferX11, switftshader use XGetWindowAttributes and
then asserts that the returned status code is zero. According to the
following manpage entry from linux.die.net:

https://linux.die.net/man/3/xgetwindowattributes

XGetWindowAttributes should return a nonzero status code on success.
This patch swaps the assert so we crash when XGetWindowAttributes fails
rather than when it succeeds.

Bug: swiftshader:168
Change-Id: Iae54f4994d0fbf8696090554a54142a4c5337a4c