bufferDeviceAddress: DeviceAddress and OpaqueCaptureAddress support
This CL ads an implementation of these functions:
- vkGetBufferDeviceAddress
- vkGetBufferOpaqueCaptureAddress
- vkGetDeviceMemoryOpaqueCaptureAddress

as well as support for OpaqueCaptureAddress in DeviceMemory
and Buffer objects. This is required by the bufferDeviceAddress
feature, which is mandatory in Vulkan 1.3.

This CL does not add support for SPV_KHR_physical_storage_buffer,
which is the other half of the bufferDeviceAddress feature, so it
can't be exposed yet.

Bug: b/184952772
Change-Id: I73ef36428ae0fd4474bc90a9e6cf8b17de0a36d9