Add a 32-bit float fma() implementation
std::fma() for float arguments may resort to using double-precision
operations (including double-precision FMA). Thus its precision may
differ from the expected binary32 IEEE 754-2008 operation. This is not
acceptable for unit testing algorithms that are expected to be
translated to efficient Reactor SIMD code.

This change uses the intrinsic for the x86 FMA instruction. Note that
this assumes we only run these tests on x86 systems, and only ones
capable of FMA operations.

Bug: b/214591655
Change-Id: I38722703e49f9a4ddd850a25d7c94d4f0c09c272