Improve the precision of highp log2()
This algorithm is based on the relaxed precision log2() algorithm, with
a higher degree polynomial and using integer subtraction to correct the
exponent bias.

Precision improved from 2.77 ULP and 0.57 margin for the absolute error,
to 1.70 and 0.36 respectively.

Compared to the previous algorithm it eliminates a division, at the cost
of three FMA operations. Latency got worse, but throughput improved, so
the effective performance impact depends on the use case and specific
CPU model characteristics.

Bug: b/169754022
Change-Id: I020203af3e70b3936b74cd2dd1d4039d274fcd9a