Improve the precision and performance of highp log2()
This algorithm is based on the relaxed precision log2() algorithm, with
a higher degree polynomial and using integer subtraction to correct the
exponent bias.

Precision improved from 2.77 ULP and 0.57 margin for the absolute error,
to 1.70 and 0.36 respectively.

It replaces an approximation based on 6 multiplications and a division,
with one using 8 multiplications. This lowers latency and increases
throughput.

Bug: b/169754022
Change-Id: I020203af3e70b3936b74cd2dd1d4039d274fcd9a