Fix handling infinity for relaxed precision log2()
Previously log2(inifity) would return 128.0f for the relaxed precision
version.

Note that this can't be handled by replacing 128.0 with infinity at the
end of the function, because 128.0 is also the correct result for finite
input values near the end of the representable range, due to rounding.

The "exhaustive" unit test has been extended to include the entire range
of valid input values.

Also some comments have been added and variables renamed.

Bug: b/169754022
Change-Id: I46cc553b8606990a11cf466f616c8eaa1e6a1cb8