Fix SPV_EXT_shader_stencil_export
The spec mentions that:
"Decorating a variable with the FragStencilRefEXT built-in
 decoration will make that variable contain the new stencil
 reference value for all samples covered by the fragment.
 This value will be used as the stencil reference value
 used in stencil testing."

Using this value is not related to using the
VK_STENCIL_OP_REPLACE mode. If the value is used in the
shader, it has to replace the stencil value, regardless
of the stencil op.

Tests: dEQP-VK.renderpass.depth_stencil_write_conditions.stencil_*
Bug: b/218397060
Change-Id: I4b5a3a834e17397ef668c3158a52b87760ca7b58