Support VK_IMAGE_CREATE_EXTENDED_USAGE_BIT
The Vulkan spec states that "For images created with
VK_IMAGE_CREATE_EXTENDED_USAGE_BIT a usage bit is valid if it is
supported for at least one of the formats a VkImageView created from the
image can have".

Previously this flag was ignored and we failed to create some images for
which the usage flag was meant for an image view with a different
format.

The Format::compatibleFormat() method was renamed to
getCompatibilityClassRepresentative(). This method is used in the
implementation of the new Format::getCompatibleFormats() method which
returns an std::vector with all the formats in the same compatibility
class.

Bug: b/214957976
Change-Id: I1db61f51744914405825c0471fee014a6b6845af