Validate SPIR-V instruction word array accesses
Previously the InsnIterator::wordPointer() method allowed unchecked
access to SPIR-V instruction words. This change eliminates this method
and instead provides a `Span` helper class to represent a contiguous
range of instruction words, where every access is checked against the
bounds of the instruction size as well as the word count of the span.

Bug: b/191426269
Change-Id: Idaef6432e8a0c489f565428edf4d5a347c6c518c