Only build WSI support for XCB surfaces when xcb.h is found
Previously we built with VK_USE_PLATFORM_XCB_KHR when the XCB library
was found, but some systems may have it without having the header
available.

This change enables VK_USE_PLATFORM_XCB_KHR when the header is found,
regardless of whether the library is found. Note that the latter still
gets dynamically checked for at run-time before we expose the
VK_KHR_xcb_surface extension.

A warning is emitted on Linux builds if xcb/xcb.h is not found since
XCB is the main WSI integration on this platform and it's typically a
build system setup bug if the libx11-xcb-dev package was not installed.

Bug: b/209030951
Change-Id: Ifa7bd23c99245025182100412d36f3d3cb3258b2