Detect <xcb/shm.h> for building XCB surface support
Building with VK_USE_PLATFORM_XCB_KHR enabled requires the <xcb/shm.h>
header. Note that we dynamically check whether the MIT-SHM extension is
actually available on the target system before making use of it, but we
need the header to be present on the build machine. A warning about
installing the libxcb-shm0-dev package gets generated if it's not.

Also, the header include order was fixed and clang-format reapplied.

Bug: b/209030951
Change-Id: I6a6889c9ee64e3b8aa2a56a15a3aeff534268638