Use 'discard' to denote terminated or demoted invocations
SPIR-V's `OpKill` instruction has been deprecated because some
implementations interpret it as fully terminating an invocation, while
others implement it as demoting an invocation to a 'helper' which still
performs arithmetic computations but doesn't perform memory writes.

This change follows suite by avoiding 'kill', and instead uses
'terminate' or 'demote-to-helper' nomenclature where the semantics need
to be explicit. Since fragment data produced by terminated or demoted
invocations must be discarded, this change uses 'discard' to denote
either of these two fates of an invocation.

Note that while HLSL uses 'discard' to specifically denote demotion to a
helper invocation, in the context of Vulkan, or elsewhere for that
matter, it does not imply whether or not arithmetic computation still
took place. So this shouldn't add any new confusion, while avoiding any
association with the chosen OpKill implementation.

Bug: b/204502919
Change-Id: I0b6a70b34a0e7a70abd74e3330e88187f095cde2