VkBufferImageCopy*: correctly detect entire rows are copied.
isEntireRow was checking that the bufferRowPitch was equal to the
image's rowPitch to know if a whole row should be copied. However
vk::Image might have extra padding on each row, which should not be
copied into the buffer (bytes not part of the copy should not be
overwritten).

Fixes the condition by checking the following condition:

    widthInBlocks * blockByteSize == bufferRowPitch

This fixes a lot of tests when running the WebGPU CTS with Swiftshader.

Bug: dawn:1347