Refactor sRGB color space conversions
We had duplicate sRGBtoLinear() functions between the Blitter and
SpirvShaderImage. The only difference was that the former only
modifies RGB components. This is now handled at the call site in the
Blitter itself, using a component mask.

This refactoring makes LinearToSRGB() available in SpirvShaderImage
for subsequent use by the OpImageWrite implementation.

Bug: b/224796232
Change-Id: I8c65a6e5b6ab7078e036f9963b2dfb9e8c25dd92