Mark Reactor routine parameters 'noundef'
Code built with the -fsanitize-memory-param-retval flag will cause
MemorySanitizer instrumentation to check function parameters before the
call, and omit updating their 'shadow' in the __msan_param_tls[] thread-
local arrays. This allows callees to assume parameters are always
initialized, but they must not check the shadow since it contains stale
data, which can lead to false positive errors.

Previously we worked around this by calling __msan_unpoison_param() to
clear the shadow in RoutineT<> calls. This can lead to false negatives
for builds without -fsanitize-memory-param-retval, and also didn't
address the issue for routine calls through a raw function pointer
instead of RoutineT<>::operator().

This change instead marks parameters as `NoUndef` for LLVM >= 13. It
assumes that if such a recent version of LLVM is being used and
detection of uninitialized parameters is desired, projects will be
compiled with -fsanitize-memory-param-retval.

Where this is not the case we're subject to the same false negatives as
before, but builds that use the integrated LLVM 10 copy regain coverage,
and this approach also works for raw function pointer calls.

Bug: b/229724343
Change-Id: I80438b9a6328c346b74c7d652661898e301777bc