Implement ScopedPragma
This change adds a ScopedPragma class which sets a pragma state and
reverts it to its old state on destruction.

For now only Boolean pragmas are supported (which in turn only has one
MemorySanitizerInstrumentation option). But since this is meant to
replace rr::Config, the implementation is based around an std::variant<>
to easily support additional settings.

PragmaBooleanOption was renamed to BooleanPragmaOption for easier
reading.

Bug: b/192879676
Change-Id: I63b9553b794b593f1bc3659e89fe3fcf3e266826