Remove use of std::iterator, which is gone in C++20.
These changes remove use in the most straightforward way possible, which
is just to add the iterator type aliases instead.  Upstream LLVM trunk
has also fixed these cases, but not always in the same way.

This also std::-qualifies one call to find() that was only working
before due to ADL on the iterator type.

Bug: chromium:1284275
Change-Id: I64445fbadada8f04c7d03738613308578f7eff4c