Don't make use of cvtps2dq in MSan builds
MemorySanitizer's instrumentation currently does not handle the cvtps2dq
intrinsic/instruction. It falls back to checking the source operand for
any uninitialized data. In shaders we can have conditional code which
causes only some SIMD lanes to be initialized, and only those lanes are
logically used in further computations. So it's valid in these cases to
do operations that use cvtps2dq on a partially initialized vector, but
MSan will report it as an error.

This can be worked around by avoiding the use of cvtps2dq and instead
relying on lowerRoundInt(), which translates into the nearbyint
intrinsic followed by an FPToSI. MemorySanitizer handles the former
in the maybeHandleSimpleNomemIntrinsic() method, which copies the shadow
of the source vector to the destination vector (i.e. it propagates it
without checking it).

Note that cvtps2dq does not follow the same code path as nearbyint
because it has different types for the source and destination vector.
We could handle it explicitly by doing the same shadow propagation.
Considering that the workaround in Reactor results in simply using two
cheap instructions instead of one it's not performance critical to fix
this in LLVM.

Bug: b/172238865
Change-Id: I4f07bb8cb6d25d914dab836f64510f8b2bad18ba