Don't make use of roundps in MSan builds
MemorySanitizer's instrumentation currently does not handle the roundps
intrinsic/instruction. It falls back to checking the source operand for
any uninitialized data. In shaders we can have conditional code which
causes only some SIMD lanes to be initialized, and only those lanes are
logically used in further computations. So it's valid in these cases to
do operations that use roundps on a partially initialized vector, but
MSan will report it as an error.

This can be worked around by avoiding the use of roundps and instead
relying on lowerRound() and lowerTrunc(), which translate into the
target-agnostic round and trunc intrinsics. MemorySanitizer handles
these in the maybeHandleSimpleNomemIntrinsic() method, which copies the
shadow of the source vector to the destination vector (i.e. it
propagates it without checking it).

Note that roudps does not follow the same code path as round and trunc
because it has an immediate operand to select the rounding mode. We
could handle it explicitly by doing the same shadow propagation.

Bug: b/172238865
Change-Id: Ib14bf4fd240b4e56462fb70429b442763f077d1e