Fix using uninitialized blend constants
Renderer::draw() unconditionally calls:
pixelProcessor.setBlendConstant(pipelineState.getBlendConstants());

which means the blend constants must be initialized before
Renderer::draw() is called. If blend constants are marked as being
dynamic in the pipeline, but vkCmdSetBlendConstants() is never
called, then blend constants are used before being initialized.

This CL adds initializers for the remaining uninitialized members of
DynamicState in order to fix this issue.

Bug: chromium:1325636
Change-Id: I940559f3fc4c0d0d52c6ebfe5ceae4706fe6906e