Optimize SPIR-V matrix operations using FMA
OpMatrixTimesVector, OpVectorTimesMatrix, OpMatrixTimesMatrix, and
OpDot each consist of dot products which can be optimized by making
use of fused multiply-add operations.

Note that the NoContraction decoration specifies that "the operation
cannot be combined with another instruction to form a single operation."
Thus it doesn't apply to the operations performed within a single
instruction. GLSL.std.450 OpFma more explicitly clarifies that
NoContraction does not affect it.

Bug: b/216472189
Change-Id: I04aa0bc2c90784cb2cfbe28034a0ec2af7fd24c9