Fix early fragment tests
zMask and sMask need to be updated whenever cMask gets modified
during the shader's execution. cMask can be modified whenever the
shader contains a discard operation, which was taken into account.
cMask can also be modified when the shader contains outputs to the
SampleMask builtin, which was NOT taken into account, so this CL
adds a proper check to see when coverage information needs to be
updated.

Tests: dEQP-VK.fragment_operations.early_fragment.sample_count_early_fragment_tests_depth_samples_4
Bug: b/200826740
Change-Id: I9f5249dc43c83b88d25ffaa9a33c8f89590d6b8d