Fix border comparison with depth 16
In SamplerCore::sampleTexel(), if we have "state.compareEnable" and
"!hasFloatTexture()", a normalization of the input value "c.x" will
occur. It is then expected that the input value should already be
scaled.

In SamplerCore::replaceBorderTexel(), we need to take this into
account and make sure the border pixels are in the same scale as the
other pixels of the same texture and output scaled borders when we
have those same conditions ("state.compareEnable" and
"!hasFloatTexture()")

Bug: b/227518327
Change-Id: Iefd5f27fbee5b7a0ca1a9eb3852ff3085e6848fe