Eliminate multiplication operators on rr::Pointer4
Multiplication of a pointer is nonsensical. This operator was abused for
signifying that the previously added offset has to be multiplied.

Instead we should just multiply the offset before adding it to the
pointer. This multiplication only took place in the InterleaveByLane()
function, so by passing it the offset we can avoid "multiplying" the
pointer. Next, since this function is only called for obtaining the
SIMD pointer to interleaved struct elements, and otherwise the offset
is uniformly added to the base pointer, this alternate code path has
been moved into this utility function, and it has been renamed to
GetElementPointer().

Bug: b/234488907
Change-Id: I67db76d5884851e86ddf1e426ae394106edde89b