createIntermediate refactor
More information will be required in order to properly create an
intermediate when pointers can also be intermediates, once the
bufferDeviceAddress feature is enabled, so this CL changes the
createIntermediate() call from a State member function call to a
SpirvShader member function call.

No new behavior is introduced by this CL, it should be a noop.

Bug: b/184952772
Change-Id: Id5b4127cae94c841c600a0cfbabef92cca406389