Avoid clamping uninitialized values when clearing depth images
The Blitter was always clamping the input clear values for
normalized formats as if they were always used for clearing
a color image, but VK_FORMAT_D16_UNORM is also a normalized
format, so this code was also reached in this case.

This CL adds a check for which aspect of the image is getting
cleared, in order to avoid clamping uninitialized values.

Bug: chromium:1336057
Change-Id: I423f14ff333382275b0882933b3b134c7980c0ba