Wrap shuffle and constant construction values
This change replaces the raw pointer of values passed to Nucleus
createConstantVector() and createShuffleVector() with an std::vector<>.
This lets us know the number of elements provided, and in turn stay
within bounds if fewer elements than the size of the vector type are
provided.

This is particularly useful for broadcast operations for arbitrary
vector sizes. Only a single element needs to be provided, and it will be
replicated to every lane. It might also prove useful for 'quad'
operations on wide vectors (i.e with more than 4 elements).

Bug: b/214583550
Change-Id: I8ddc0d8037f1c6e682c58c605682ad1f752cb9c0