Set conformance version to an acceptable value
Since we now expose Vulkan 1.3, the acceptable values for conformance
version, according to
https://gitlab.khronos.org/Tracker/vk-gl-cts/-/blob/main/external/vulkancts/modules/vulkan/api/vktApiDriverPropertiesTests.cpp#L57,
are currently:
- 1, 3, 3, 0
- 1, 3, 2, 0
- 1, 3, 1, 1
- 1, 3, 1, 0
- 1, 3, 0, 0

We've been testing against main, which is 1.3.2.0, so let's use that
for now. We can always change it later if we want to target a
different version, but we're fairly close to achieving conformance
for 1.3.3.0, so might as well target it.

Tests: dEQP-VK.api.driver_properties.conformance_version
Bug: b/204169571
Change-Id: Icebb92b588963517beba5653df75e6c0b0a0ec3e